Authorization

+++ routes/posts.js
@@ -39,7 +39,8 @@ router.post("",
         const post = new Post({
             title: req.body.title,
             Content: req.body.Content,
-            imagePath: url + "/images/" + req.file.filename
+            imagePath: url + "/images/" + req.file.filename,
+            creator: req.userData.userId
         });
         post.save().then(
             createdPost => {
@@ -74,16 +75,24 @@ router.put("/:id",
             _id: req.body.id,
             title: req.body.title,
             Content: req.body.Content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
-        Post.updateOne({_id: req.params.id}, post).then(
+        Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(
             result => {
-                console.log(result)
-                res.status(200).json(
-                    {
-                        message: "Post Updated Successfully!",
-                    }
-                );
+                if (result.nModified > 0) {
+                    res.status(200).json(
+                        {
+                            message: "Post Updated Successfully!",
+                        }
+                    );
+                } else {
+                    res.status(401).json(
+                        {
+                            message: "Unauthorised Request",
+                        }
+                    );
+                }
             }
         );
     }
@@ -140,13 +149,21 @@ router.get("/:id",
 router.delete("/:id",
     checkAuth,
     (req, res, next) => {
-        Post.deleteOne({_id: req.params.id}).then(
+        Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(
             result => {
-                console.log(result);
+            if (result.n > 0) {
+                res.status(200).json(
+                    {
+                        message: "Deletion Successfully!",
+                    }
+                );
+            } else {    
+                res.status(401).json(
+                    {
+                        message: "Unauthorised Request",
+                    }
+                );
             }
-        )
-        res.status(200).json({
-            message: "Post deleted"
         });
     }
 );
